+++ RELEASE-NOTES.md
@@ -1,6 +1,7 @@
 Release 3.2
 
 * #405 Upgrade to Cayenne 4.0.1
+* #406 Overlapping relationship / attribute includes are order dependent
 * #407 Stop using Include / Exclude as dummy containers of other Includes / Excludes
 * #408 Change parameter injection to Strings / Integers